+++ Cross.CQRS/Commands/Command`.cs
@@ -3,7 +3,7 @@
 /// <summary>
 /// Base implementation of <see cref="ICommand{TResult}"/>.
 /// </summary>
-public abstract class Command<TResult> : ICommand<TResult>
+public abstract record Command<TResult> : ICommand<TResult>
 {
     /// <inheritdoc />
     public Guid CommandId { get; } = Guid.NewGuid();