+++ Cross.CQRS/Events/CommandEvent.cs
@@ -1,0 +1,18 @@
+﻿namespace Cross.CQRS.Events;
+
+/// <summary>
+/// Base implementation of <see cref="ICommandEvent"/>.
+/// </summary>
+public abstract record CommandEvent : ICommandEvent
+{
+    protected CommandEvent(Guid commandId)
+    {
+        CommandId = commandId;
+    }
+
+    /// <inheritdoc />
+    public Guid CommandId { get; }
+
+    /// <inheritdoc />
+    public virtual CommandEventFlowTypeEnum EventFlowType() => CommandEventFlowTypeEnum.StandardFlow;
+}